fix: use std::size_t for vertex IDs to support x86

+++ OpenGL/geometry/half_edge_mesh.h
@@ -21,7 +21,7 @@ namespace geometry {
 		[[nodiscard]] const auto& Vertices() const { return vertices_; }
 		[[nodiscard]] const auto& Edges() const { return edges_; }
 		[[nodiscard]] const auto& Faces() const { return faces_; }
-		[[nodiscard]] std::uint64_t NextVertexId() { return next_vertex_id_++; }
+		[[nodiscard]] std::size_t NextVertexId() { return next_vertex_id_++; }
 
 		operator gfx::Mesh() const;
 
@@ -32,6 +32,6 @@ namespace geometry {
 		std::map<std::size_t, std::shared_ptr<Vertex>> vertices_;
 		std::unordered_map<std::size_t, std::shared_ptr<HalfEdge>> edges_;
 		std::unordered_map<std::size_t, std::shared_ptr<Face>> faces_;
-		std::uint64_t next_vertex_id_;
+		std::size_t next_vertex_id_;
 	};
 }
